refactor: rename/extract classes
