+++ packages/client/src/Config.ts
@@ -6,7 +6,7 @@ import addFormats from 'ajv-formats'
 import merge from 'lodash/merge'
 
 import type { AuthConfig, EthereumConfig } from './Ethereum'
-import type { EncryptionConfig } from './encryption/KeyExchangeUtils'
+import type { EncryptionConfig } from './encryption/KeyExchangeStream'
 
 import CONFIG_SCHEMA from './config.schema.json'
 import { EthereumAddress, SmartContractRecord } from 'streamr-client-protocol'
