+++ packages/client/src/publish/PublishPipeline.ts
@@ -7,7 +7,8 @@ import { scoped, Lifecycle, inject, delay } from 'tsyringe'
 import { inspect } from '../utils/log'
 import { instanceId, Defer, Deferred } from '../utils'
 import { Context, ContextError } from '../utils/Context'
-import { PushPipeline, Pipeline } from '../utils/Pipeline'
+import { Pipeline } from '../utils/Pipeline'
+import { PushPipeline } from '../utils/PushPipeline'
 import { Stoppable } from '../utils/Stoppable'
 
 import { MessageCreator } from './MessageCreator'
