+++ packages/client/src/publish/Publisher.ts
@@ -7,9 +7,10 @@ import { scoped, Lifecycle, inject, delay } from 'tsyringe'
 import { instanceId } from '../utils'
 import { Context } from '../utils/Context'
 import { CancelableGenerator, ICancelable } from '../utils/iterators'
-import { PublishPipeline, PublishMetadata } from './PublishPipeline'
+
+import { PublishMetadata, PublishPipeline } from './PublishPipeline'
 import { Stoppable } from '../utils/Stoppable'
-import { PublisherKeyExchange } from '../encryption/KeyExchangePublisher'
+import { PublisherKeyExchange } from '../encryption/PublisherKeyExchange'
 import { StreamDefinition } from '../types'
 
 export type { PublishMetadata }
