+++ packages/client/test/unit/GroupKeyPersistence.test.ts
@@ -2,7 +2,7 @@ import crypto from 'crypto'
 import fs from 'fs'
 import { join } from 'path'
 import { tmpdir } from 'os'
-import { GroupKey } from '../../src/encryption/Encryption'
+import { GroupKey } from '../../src/encryption/GroupKey'
 import { GroupKeyPersistence } from '../../src/encryption/GroupKeyStore'
 import { uid, addAfterFn, describeRepeats, mockContext } from '../test-utils/utils'
 import LeakDetector from 'jest-leak-detector'
