+++ packages/client/test/unit/waitForAssignmentsToPropagate.test.ts
@@ -1,6 +1,6 @@
 import { waitForAssignmentsToPropagate } from '../../src/utils/waitForAssignmentsToPropagate'
 import { MessageID, StreamID, StreamMessage, StreamPartID, toStreamID, toStreamPartID } from 'streamr-client-protocol'
-import { PushPipeline } from '../../src/utils/Pipeline'
+import { PushPipeline } from '../../src/utils/PushPipeline'
 import { range, shuffle } from 'lodash'
 import { wait } from 'streamr-test-utils'
